+++ lib/features/general/custom_chip.widget.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
 class CustomChip extends StatelessWidget {
   final Text label;
@@ -25,7 +24,7 @@ class CustomChip extends StatelessWidget {
       child: ClipRRect(
         borderRadius: BorderRadius.circular(10),
         child: Container(
-          color: color ?? Get.theme.chipTheme.backgroundColor,
+          color: color ?? Theme.of(context).chipTheme.backgroundColor,
           padding: padding,
           child: label,
         ),